+++ src/pages/MainEditor/components/SideBar/components/AddDialogButton.tsx
@@ -1,5 +1,6 @@
 import AddDialogModalStore from "../../../Modals/Modal_AddDialog/AddDialogModalStore";
 import "../../RightSideBar/components/AddButton.css"
+// FIXME: исправить юзание css файла с другой папки
 
 export default function () {
   return (
